\name{OjaNP-package}
\alias{OjaNP-package}
\docType{package}
\title{
Multivariate Methods Based on the Oja Median and Related Concepts 
}
\description{
The package provides functions for the Oja median, Oja signs and ranks and methods based upon them.
}
\details{
\tabular{ll}{
Package: \tab OjaNP\cr
Type: \tab Package\cr
Version: \tab 0.9-9\cr
Date: \tab 2016-04-24\cr
License: \tab GPL (>= 2)\cr
LazyLoad: \tab yes\cr
}
The package implements several algorithms to compute the Oja median as well as the Oja signs and ranks and methods based upon them.
}
\author{
Daniel Fischer, Karl Mosler, Jyrki \enc{Möttönen}{Mottonen}, Klaus Nordhausen, Oleksii Pokotylo,  Daniel Vogel.

Maintainer: Daniel Fischer <daniel.fischer@luke.fi>

}

\keyword{ package }

